# contracts/host/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisNoteScope": {
        "anyOf": [
          {
            "additionalProperties": false,
            "properties": {
              "kind": {
                "enum": [
                  "global"
                ],
                "type": "string"
              }
            },
            "required": [
              "kind"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "contextId": {
                "pattern": "^[1-9]\\d{0,15}$",
                "type": "string"
              },
              "kind": {
                "enum": [
                  "context"
                ],
                "type": "string"
              }
            },
            "required": [
              "kind",
              "contextId"
            ],
            "type": "object"
          }
        ]
      },
...
      "CreateAnalysisNoteResult": {
        "additionalProperties": false,
        "properties": {
          "anchorId": {
            "pattern": "^[1-9]\\d{0,15}$",
            "type": "string"
          },
          "contextId": {
            "pattern": "^[1-9]\\d{0,15}$",
            "type": "string"
          },
          "contributionId": {
            "pattern": "^[1-9]\\d{0,15}$",
            "type": "string"
          },
          "dataRevision": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "itemId": {
            "pattern": "^[1-9]\\d{0,15}$",
            "type": "string"
          },
          "resumeUpdate": {
            "additionalProperties": false,
            "properties": {
              "contextId": {
                "pattern": "^[1-9]\\d{0,15}$",
                "type": "string"
              },
              "resumeVersion": {
                "maximum": 9007199254740991,
                "minimum": 1,
                "type": "integer"
              }
            },
            "required": [
              "contextId",
              "resumeVersion"
            ],
            "type": "object"
          },
          "revisionId": {
            "pattern": "^[1-9]\\d{0,15}$",
            "type": "string"
          },
          "scopeKind": {
            "anyOf": [
              {
                "enum": [
                  "global"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "context"
                ],
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "contributionId",
          "itemId",
          "revisionId",
          "anchorId",
          "scopeKind",
          "dataRevision"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "correlationId": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "parameters": {
            "additionalProperties": false,
            "properties": {
              "currentRevision": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              },
              "expectedRevision": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              }
            },
            "type": "object"
          },
          "retryable": {
            "type": "boolean"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "correlationId",
          "retryable",
          "parameters"
        ],
        "type": "object"
      },
...
      "WorkScope": {
        "anyOf": [
          {
            "additionalProperties": false,
            "properties": {
              "kind": {
                "enum": [
                  "free"
                ],
                "type": "string"
              }
            },
            "required": [
              "kind"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "contextId": {
                "pattern": "^[1-9]\\d{0,15}$",
                "type": "string"
              },
              "kind": {
                "enum": [
                  "context"
                ],
                "type": "string"
              }
            },
            "required": [
              "kind",
              "contextId"
            ],
            "type": "object"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "hostBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Plysmith Application Host",
    "version": "0.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/analysis/notes": {
      "post": {
        "operationId": "CreateAnalysisNote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "expectedScratchId": {
                    "maxLength": 160,
                    "minLength": 1,
                    "pattern": "^[A-Za-z0-9._:-]+$",
                    "type": "string"
                  },
                  "expectedScratchRevision": {
                    "maximum": 9007199254740991,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "languageTag": {
                    "maxLength": 35,
                    "minLength": 1,
                    "type": "string"
                  },
                  "noteScope": {
                    "$ref": "#/components/schemas/AnalysisNoteScope"
                  },
                  "scope": {
                    "$ref": "#/components/schemas/WorkScope"
                  }
                },
                "required": [
                  "scope",
                  "expectedScratchId",
                  "expectedScratchRevision",
                  "languageTag",
                  "noteScope"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAnalysisNoteResult"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    },
    "/analysis/notes/{contributionId}": {
      "delete": {
        "operationId": "DeleteAnalysisNote",
        "parameters": [
          {
            "in": "path",
            "name": "contributionId",
            "required": true,
            "schema": {
              "pattern": "^[1-9]\\d{0,15}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "expectedContributionVersion": {
                    "maximum": 9007199254740991,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "scope": {
                    "$ref": "#/components/schemas/WorkScope"
                  }
                },
                "required": [
                  "scope",
                  "expectedContributionVersion"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisNoteMutationResult"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      },
      "patch": {
        "operationId": "UpdateAnalysisNote",
        "parameters": [
          {
            "in": "path",
            "name": "contributionId",
            "required": true,
            "schema": {
              "pattern": "^[1-9]\\d{0,15}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "body": {
                    "maxLength": 8000,
                    "minLength": 1,
                    "type": "string"
                  },
                  "expectedContributionVersion": {
                    "maximum": 9007199254740991,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "scope": {
                    "$ref": "#/components/schemas/WorkScope"
                  }
                },
                "required": [
                  "scope",
                  "expectedContributionVersion",
                  "body"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisNoteMutationResult"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    },
    "/analysis/position-notes": {
      "post": {
        "operationId": "CreatePositionNote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "anchorId": {
                    "pattern": "^[1-9]\\d{0,15}$",
                    "type": "string"
                  },
                  "body": {
                    "maxLength": 8000,
                    "minLength": 1,
                    "type": "string"
                  },
                  "itemId": {
                    "pattern": "^[1-9]\\d{0,15}$",
                    "type": "string"
                  },
                  "languageTag": {
                    "maxLength": 35,
                    "minLength": 1,
                    "type": "string"
                  },
                  "noteScope": {
                    "$ref": "#/components/schemas/AnalysisNoteScope"
                  },
                  "revisionId": {
                    "pattern": "^[1-9]\\d{0,15}$",
                    "type": "string"
                  },
                  "scope": {
                    "$ref": "#/components/schemas/WorkScope"
                  }
                },
                "required": [
                  "scope",
                  "itemId",
                  "revisionId",
                  "anchorId",
                  "body",
                  "languageTag",
                  "noteScope"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisNoteMutationResult"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    },
...
    "/analysis/workspace": {
      "get": {
        "operationId": "GetAnalysisWorkspace",
        "parameters": [
          {
            "in": "query",
            "name": "scopeKind",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "free"
                  ],
                  "type": "string"
                },
                {
                  "enum": [
                    "context"
                  ],
                  "type": "string"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "contextId",
            "required": false,
            "schema": {
              "pattern": "^[1-9]\\d{0,15}$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "itemId",
            "required": false,
            "schema": {
              "pattern": "^[1-9]\\d{0,15}$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revisionId",
            "required": false,
            "schema": {
              "pattern": "^[1-9]\\d{0,15}$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "anchorId",
            "required": false,
            "schema": {
              "pattern": "^[1-9]\\d{0,15}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisWorkspace"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    },
...
  },
  "security": [
    {
      "hostBearer": []
    }
  ]
}